Move getLinodeDescription to utils
